refactor to use server actions instead of routes
