--- src/app/api/forecast/route.ts
@@ -1,7 +1,0 @@
-import { fetchWeather } from "@/app/utils/fetch-weather";
-import { NextRequest } from "next/server";
-
-export async function PUT(_: NextRequest) {
-  const data = await fetchWeather();
-  return Response.json(data);
-}
