+++ src/app/utils/day-of-week.ts
@@ -1,12 +1,4 @@
-const WEEK_DAYS = ["日", "月", "火", "水", "木", "金", "土"] as const;
-const FUTURE_DAYS_LOOKUP: Record<number, string> = {
-  0: "今日",
-  1: "明日",
-  2: "明後日",
-  3: "３日後",
-  4: "４日後",
-  5: "５日後",
-} as const;
+import { FUTURE_DAYS_LOOKUP, WEEK_DAYS } from "../constants/japan-dates";
 
 export function dayOfWeekInJapanese(date: Date) {
   const futureDays = Object.keys(FUTURE_DAYS_LOOKUP).find((day) =>
